Prevent empty space before `inout`
When `inout` is used it should follow a node which appends a space so it
does not need to prepend a space itself (we still need to append a space
to separate it from the following identifier). Not prepending a space
here avoids inserting extra spaces if e.g., `inout` is the token in a
function argument list.